+++ packages/author-profile/author-profile-content.web.js
@@ -157,13 +157,10 @@ class AuthorProfileContent extends Component {
       refetch
     } = this.props;
 
-    const paginationComponent = (
-      { hideResults = false, hideTopKeyline = false } = {}
-    ) => (
+    const paginationComponent = ({ hideResults = false } = {}) => (
       <AuthorProfilePagination
         count={count}
         hideResults={hideResults}
-        hideTopKeyline={hideTopKeyline}
         onNext={(...args) => {
           onNext(...args);
           scrollUpToPaging();
@@ -198,7 +195,7 @@ class AuthorProfileContent extends Component {
 
     const Contents = (
       <ContentContainer>
-        {paginationComponent({ hideResults: false, hideTopKeyline: true })}
+        {paginationComponent({ hideResults: false })}
         <View style={styles.container}>
           {data &&
             data.map((article, key) => {