+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.10.3"></a>
+## [0.10.3](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.10.2...@times-components/article-flag@0.10.3) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
+<a name="0.10.2"></a>
+## [0.10.2](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.10.0...@times-components/article-flag@0.10.2) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
+<a name="0.10.1"></a>
+## [0.10.1](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.10.0...@times-components/article-flag@0.10.1) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
 <a name="0.10.0"></a>
 # [0.10.0](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.9.10...@times-components/article-flag@0.10.0) (2018-01-18)