+++ packages/image/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.12.8"></a>
+## [1.12.8](https://github.com/newsuk/times-components/compare/@times-components/image@1.12.7...@times-components/image@1.12.8) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
+<a name="1.12.7"></a>
+## [1.12.7](https://github.com/newsuk/times-components/compare/@times-components/image@1.12.5...@times-components/image@1.12.7) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
+<a name="1.12.6"></a>
+## [1.12.6](https://github.com/newsuk/times-components/compare/@times-components/image@1.12.5...@times-components/image@1.12.6) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
 <a name="1.12.5"></a>
 ## [1.12.5](https://github.com/newsuk/times-components/compare/@times-components/image@1.12.4...@times-components/image@1.12.5) (2018-01-18)