+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.6"></a>
+## [0.6.6](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.6.5...@times-components/tracking@0.6.6) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="0.6.5"></a>
+## [0.6.5](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.6.3...@times-components/tracking@0.6.5) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.6.3...@times-components/tracking@0.6.4) (2018-01-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
 <a name="0.6.3"></a>
 ## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.6.2...@times-components/tracking@0.6.3) (2018-01-18)